requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.53.0
accelerate>=0.30.0
qwen-vl-utils>=0.0.8
pillow>=10.0.0
tqdm>=4.60.0
